requirements.txt
# SentinelHunter dependencies
# Run `pip freeze > requirements.txt` in production to lock exact versions.

opensearch-py>=2.0.0,<3.0.0
networkx>=3.0,<4.0
torch>=2.0.0,<3.0.0
torch-geometric>=2.3.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.1.0
requests>=2.31.0,<3.0.0
pycti>=6.0.0
transformers>=4.35.0
pyod>=1.1.0,<2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
streamlit>=1.28.0
pyvis>=0.3.0
google-generativeai>=0.7.0
joblib>=1.3.0,<2.0.0
python-dateutil>=2.8.0
